+++ src/pages/levelCreator/CodeModal.tsx
@@ -10,6 +10,7 @@ import InfoOutlinedIcon from "@mui/icons-material/InfoOutlined"
 import InputLabel from "@mui/material/InputLabel"
 import MenuItem from "@mui/material/MenuItem"
 import Modal from "@mui/material/Modal"
+import TextField from "@mui/material/TextField"
 import Tooltip from "@mui/material/Tooltip"
 import Typography from "@mui/material/Typography"
 
@@ -72,6 +73,8 @@ const CodeModal: FC<CodeModalProps> = ({ open, onClose }) => {
   const handleChange = (event: SelectChangeEvent) => {
     setLanguage(event.target.value)
   }
+
+  const [maxMoves, setMaxMoves] = useState(50)
 
   const [blockCounts, setBlockCounts] = useState<Record<string, BlockCount>>(
     () => Object.fromEntries(BLOCKS.map(block => [block.type, "infinite"])),
@@ -141,7 +144,7 @@ const CodeModal: FC<CodeModalProps> = ({ open, onClose }) => {
         <Typography>
           Here you can select the code you can use while playing your new level!
         </Typography>
-        <Box sx={{ mb: 2 }}>
+        <Box sx={{ mb: 2, display: "flex", gap: 2, alignItems: "flex-start" }}>
           <FormControl>
             <InputLabel id="language-label">Language</InputLabel>
             <Select
@@ -168,6 +171,27 @@ const CodeModal: FC<CodeModalProps> = ({ open, onClose }) => {
               ))}
             </Select>
           </FormControl>
+          <TextField
+            type="number"
+            label="Max steps"
+            variant="outlined"
+            size="medium"
+            value={maxMoves}
+            onChange={event => {
+              const value = Number(event.target.value)
+              if (!Number.isNaN(value)) {
+                setMaxMoves(Math.min(100, Math.max(1, value)))
+              }
+            }}
+            slotProps={{ htmlInput: { min: 1, max: 100, size: 4 } }}
+            sx={{
+              width: "fit-content",
+              // The shared theme forces a black border onto the filled
+              // variant's root, which conflicts with this field's outlined
+              // fieldset border, so it needs to be removed here.
+              "& .MuiOutlinedInput-root": { border: "none !important" },
+            }}
+          />
         </Box>
         {language !== "Python" && (
           <>
